[AD-303] Ensure to close AutoCloseable classes. https://bitquill.atlassian.net/browse/AD-303
