+++ calcite-adapter/src/main/java/software/amazon/documentdb/jdbc/persist/FileSchemaWriter.java
@@ -149,4 +149,9 @@ public class FileSchemaWriter implements SchemaWriter {
             }
         }
     }
+
+    @Override
+    public void close() throws Exception {
+
+    }
 }
