+++ calcite-adapter/src/test/java/software/amazon/documentdb/jdbc/query/DocumentDbQueryMappingServiceDateTimeTest.java
@@ -91,9 +91,10 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
     }
 
     @AfterAll
-    static void afterAll() throws SQLException {
-        final SchemaWriter schemaWriter = SchemaStoreFactory.createWriter(connectionProperties, client);
-        schemaWriter.remove("id");
+    static void afterAll() throws Exception {
+        try (SchemaWriter schemaWriter = SchemaStoreFactory.createWriter(connectionProperties, client)) {
+            schemaWriter.remove("id");
+        }
         client.close();
     }
