+++ src/test/java/software/amazon/documentdb/jdbc/DocumentDbResultSetMetaDataTest.java
@@ -64,13 +64,14 @@ public class DocumentDbResultSetMetaDataTest extends DocumentDbFlapDoodleTest {
     }
 
     @AfterEach
-    void afterEach() throws SQLException {
+    void afterEach() throws Exception {
         final DocumentDbConnectionProperties properties = DocumentDbConnectionProperties
                 .getPropertiesFromConnectionString(new Properties(),
                         getJdbcConnectionString(),
                 "jdbc:documentdb:");
-        final SchemaWriter schemaWriter = SchemaStoreFactory.createWriter(properties, null);
-        schemaWriter.remove(DocumentDbSchema.DEFAULT_SCHEMA_NAME);
+        try (SchemaWriter schemaWriter = SchemaStoreFactory.createWriter(properties, null)) {
+            schemaWriter.remove(DocumentDbSchema.DEFAULT_SCHEMA_NAME);
+        }
     }
 
     /**
