+++ src/test/java/software/amazon/documentdb/jdbc/DocumentDbResultSetTest.java
@@ -96,15 +96,16 @@ public class DocumentDbResultSetTest extends DocumentDbFlapDoodleTest {
     }
 
     @AfterEach
-    void afterEach() throws SQLException {
+    void afterEach() throws Exception {
         final DocumentDbConnectionProperties properties = DocumentDbConnectionProperties
                 .getPropertiesFromConnectionString(
                         new Properties(),
                         getJdbcConnectionString(),
                         "jdbc:documentdb:");
 
-        final SchemaWriter schemaWriter = SchemaStoreFactory.createWriter(properties,  null);
-        schemaWriter.remove(DocumentDbSchema.DEFAULT_SCHEMA_NAME);
+        try (SchemaWriter schemaWriter = SchemaStoreFactory.createWriter(properties,  null)) {
+            schemaWriter.remove(DocumentDbSchema.DEFAULT_SCHEMA_NAME);
+        }
     }
 
     @AfterAll
